- Optimize fragments list handling.

+++ cube/patches/base/frag.h
@@ -21,6 +21,7 @@
 #define FRAG_H
 
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 
 typedef struct {
@@ -37,7 +38,7 @@ int do_read(void *buffer, uint32_t length, uint32_t offset, uint32_t sector);
 int do_write(void *buffer, uint32_t length, uint32_t offset, uint32_t sector);
 void end_read(void);
 
-bool is_frag_patch(uint32_t offset, uint32_t size);
+bool is_frag_patch(uint32_t offset, size_t size);
 
 bool frag_read_async(void *buffer, uint32_t length, uint32_t offset, frag_read_cb callback);
 void frag_read_complete(void *buffer, uint32_t length, uint32_t offset);
